+++ crates/gpui/src/platform/keystroke.rs
@@ -7,9 +7,7 @@ use std::{
     ops::BitOr,
 };
 
-use crate::chars_for_modified_key;
-
-use super::{KeyCode, KeyboardMapper};
+use crate::{KeyCode, KeyboardMapper};
 
 /// A keystroke and associated metadata generated by the platform
 #[derive(Clone, Debug, Eq, PartialEq, Default, Hash)]
